Move the Monocle EPD tests to the system tests
Move the Monocle EPD unit tests to the system test directory where all
the other Monocle tests are located. Run the tests with the command:

    gradle -PFULL_TEST=true :systemTests:test \
        --tests test.com.sun.glass.ui.monocle.*

EPDSettingsTest: Print a notice of which error log messages are
expected.

EPDSettingsTest: Put the method parameters for Assert.assertEquals in
the correct order (expected, actual).

FramebufferY8Test: Use the @Ignore annotation instead of a boolean value
to skip tests.

FramebufferY8Test: Print a warning message when the method in the
subclass is slower than the one it overrides in the superclass for the
16-bit RGB565 conversions.