Work around problems on the Kobo Glo HD Model N437
EPDFrameBuffer

Ignore the error ENOTTY (25), "Inappropriate ioctl for device," when
setting the waveform modes. The IOCTL call is ignored by the driver on
the Kobo Glo HD where the error occurs, anyway.

Use the number of bytes per pixel (Integer.BYTES), not the number of
bits per pixel (Integer.SIZE), when calculating the capacity of the
32-bit off-screen byte buffer.

Use the visible y-resolution (yres), not the virtual y-resolution
(yresVirtual), when calculating the capacity of the off-screen byte
buffer and the length of the frame buffer mapping. The reported virtual
y-resolution can be incorrect.

Log the mapping and unmapping of the frame buffer. Log any errors that
occur on either of the system calls.

Make a note in the comments about the error that occurs when using an
8-bit frame buffer on the Kobo Clara HD Model N249 (marked as a TODO
action item).

EPDScreen

Handle a null return value from EPDFrameBuffer.getMappedBuffer.

EPDSystem

Replace the comments about passing an integer in an IOCTL call with the
correct descriptions.

FramebufferY8

Delete the extra blank line at the end of each switch block.