Rename the new system property and add notes
Rename the new system property to "monocle.epd.fixKoboN249" so that it's
specific to working around the frame buffer bug on the Kobo Clara HD
(Model N249) and should be used only when the bug is present on the
device.

Add implementation notes that explain how the names of the system
properties correspond to fields and constants in the Linux system
headers.