Rename system property to monocle.epd.fixWidthY8UR
Rename the system property to "monocle.epd.fixWidthY8UR" so that it's
specific to the work-around required instead of referring to a device
model. The bug might occur on other models besides the Kobo N249, or the
bug might exist on just the one device that was tested.

Add a debug log message that prints the Monocle native screen resolution
and color depth.